requirements.txt
asgiref==3.7.2
Django==4.2.7
django-tinymce==3.6.1
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
